fix rrdtool 1.5.x compatibility
in version 1.5.x the --showtime option was removed, thus no timestamps
in the row elements.
But we can easily calculate them using the start timestamp and the step
size.

That option seems to be back in rrdtool 1.6 but nobody knows in
what lts distribution 1.5 will land, this soloution should work
regardless of the of the version, no checking necessary.